fix(sidebar separation): Restructured files
