+++ src/internal/function.go
@@ -16,7 +16,9 @@ import (
 	"strings"
 	"time"
 
-	"github.com/yorukot/superfile/src/internal/common/utils"
+	"github.com/yorukot/superfile/src/internal/ui/sidebar"
+
+	"github.com/yorukot/superfile/src/internal/utils"
 
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/yorukot/superfile/src/internal/common"
@@ -179,7 +181,7 @@ func returnDirElementBySearchString(location string, displayDotFile bool, search
 	}
 	// https://github.com/reinhrst/fzf-lib/blob/main/core.go#L43
 	// No sorting needed. fzf.DefaultOptions() already return values ordered on Score
-	fzfResults := common.FzfSearch(searchString, fileAndDirectories)
+	fzfResults := sidebar.FzfSearch(searchString, fileAndDirectories)
 	dirElement := make([]element, 0, len(fzfResults))
 	for _, item := range fzfResults {
 		resultItem := folderElementMap[item.Key]
