+++ src/internal/ui/sidebar/consts.go
@@ -1,4 +1,4 @@
-package common
+package sidebar
 
 // Todo , merge this and predefined variables file
 // These are effectively consts
@@ -15,5 +15,3 @@ var DiskDividerDir = Directory{ //nolint: gochecknoglobals // This is more like 
 
 // superfile logo + blank line + search bar
 const SideBarInitialHeight = 3
-
-const InvalidTypeString = "InvalidType"
